Fixed all garbage code
