+++ server/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "ts-node-dev --respawn --pretty --transpile-only server/src/index.ts",
+    "dev": "ts-node-dev --respawn --pretty --transpile-only src/index.ts",
     "test": "jest"
   },
   "author": "",
