+++ server/src/start.ts
@@ -1,7 +1,9 @@
 import express from "express";
-import { router } from "./router/projects";
+import { projectRouter } from "./router/project-router";
+import { userRouter } from "./router/user-router";
 
 export const app = express();
 
 app.use(express.json());
-app.use("/projects", router);
+app.use("/projects", projectRouter);
+app.use("/users", userRouter);
